requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3

# Database
psycopg2-binary==2.9.9

# Cache & Async Tasks
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# Authentication
django-allauth==0.57.0
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0

# Image Processing
Pillow==10.1.0

# Security
argon2-cffi==23.1.0

# CORS
django-cors-headers==4.3.0

# Environment variables
python-decouple==3.8

# Development
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
django-extensions==3.2.3

# Forms & Validation
django-crispy-forms==2.1
crispy-tailwind==0.5.0
